requirements_mcp.txt
# MCP (Model Context Protocol) Dependencies
mcp>=0.1.0
psutil>=5.9.0
pyautogui>=0.9.54

# Optional but recommended
Pillow>=10.0.0  # For screenshot support
